server/requirements.txt
alembic==1.14.1
aniso8601==10.0.1
asttokens==3.0.0
backcall==0.2.0
blinker==1.9.0
click==8.1.8
decorator==5.2.1
executing==2.2.1
Faker==35.2.2
Flask==2.2.5
Flask-Cors==5.0.0
Flask-Migrate==4.1.0
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.3
greenlet==3.1.1
importlib_metadata==8.5.0
importlib_resources==6.4.5
ipdb==0.13.9
ipython==8.12.3
itsdangerous==2.2.0
jedi==0.19.2
Jinja2==3.1.6
Mako==1.3.10
MarkupSafe==2.1.5
matplotlib-inline==0.1.7
parso==0.8.5
pexpect==4.9.0
pickleshare==0.7.5
prompt_toolkit==3.0.52
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
python-dateutil==2.9.0.post0
pytz==2025.2
setuptools==70.3.0
six==1.17.0
SQLAlchemy==2.0.43
SQLAlchemy-serializer==1.4.12
stack-data==0.6.3
toml==0.10.2
traitlets==5.14.3
typing_extensions==4.13.2
wcwidth==0.2.14
Werkzeug==2.2.2
zipp==3.20.2
Flask-Bcrypt==1.0.1
PyJWT==2.8.0
